+++ src/divstats-winstats.h
@@ -41,8 +41,7 @@ double pi_k2(HaplotypeFrequencySpectrum *hfs, int k, pair_t *subset_snps = NULL)
 double pi_numerator_btw_pools(string *haps1, int length1, string *haps2, int length2, map<string, int> &hap2count, pair_t *subset_snps = NULL);
 double pi_numerator(string *haps, int length, map<string, int> &hap2count, pair_t *subset_snps = NULL);
 int hamming_dist_str(string one, string two, pair_t *subset_snps = NULL);
-int hamming_dist_ptr(short *one, short *two, int length);
-int hamming_dist_ptr(char *one, char *two, int length);
+int hamming_dist_packed(const char *one, const char *two, long start, int length);
 
 double tajimaD_from_sfs(array_t *sfs, double pi = -9, double S = -9);
 double thetaH_from_sfs(array_t *sfs);